split test files according to class under test
